backend/requirements.txt
flask>=2.3.2
flask-cors>=3.0.10
tensorflow>=2.13.0  # or 2.15.0 (latest stable)
pillow>=9.5.0
numpy>=1.24.3